+++ app/src/main/java/com/techcourse/controller/LoginController.java
@@ -5,8 +5,8 @@ import com.interface21.web.bind.annotation.RequestMapping;
 import com.interface21.web.bind.annotation.RequestMethod;
 import com.interface21.webmvc.servlet.ModelAndView;
 import com.interface21.webmvc.servlet.view.JspView;
-import com.techcourse.BadRequestException;
-import com.techcourse.UnauthorizedException;
+import com.interface21.webmvc.servlet.mvc.exception.BadRequestException;
+import com.interface21.webmvc.servlet.mvc.exception.UnauthorizedException;
 import com.techcourse.domain.User;
 import com.techcourse.repository.InMemoryUserRepository;
 import jakarta.servlet.http.HttpServletRequest;
