+++ app/src/main/java/com/techcourse/controller/RegisterController.java
@@ -5,7 +5,7 @@ import com.interface21.web.bind.annotation.RequestMapping;
 import com.interface21.web.bind.annotation.RequestMethod;
 import com.interface21.webmvc.servlet.ModelAndView;
 import com.interface21.webmvc.servlet.view.JspView;
-import com.techcourse.BadRequestException;
+import com.interface21.webmvc.servlet.mvc.exception.BadRequestException;
 import com.techcourse.domain.User;
 import com.techcourse.repository.InMemoryUserRepository;
 import jakarta.servlet.http.HttpServletRequest;
