+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/HandlerAdapterRegistry.java
@@ -1,7 +1,5 @@
-package com.techcourse;
+package com.interface21.webmvc.servlet.mvc;
 
-import com.interface21.webmvc.servlet.mvc.HandlerAdapter;
-import com.interface21.webmvc.servlet.mvc.tobe.AnnotationHandlerAdapter;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
@@ -9,10 +7,6 @@ import java.util.Optional;
 public class HandlerAdapterRegistry {
 
     private final List<HandlerAdapter> handlerAdapters = new ArrayList<>();
-
-    public void initialize() {
-        addHandlerAdapter(new AnnotationHandlerAdapter());
-    }
 
     public void addHandlerAdapter(final HandlerAdapter handlerAdapter) {
         this.handlerAdapters.add(handlerAdapter);
