+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/HandlerMappingRegistry.java
@@ -1,7 +1,5 @@
-package com.techcourse;
+package com.interface21.webmvc.servlet.mvc;
 
-import com.interface21.webmvc.servlet.mvc.HandlerMapping;
-import com.interface21.webmvc.servlet.mvc.tobe.AnnotationHandlerMapping;
 import jakarta.servlet.http.HttpServletRequest;
 import java.util.ArrayList;
 import java.util.List;
@@ -10,13 +8,6 @@ import java.util.Optional;
 public class HandlerMappingRegistry {
 
     private final List<HandlerMapping> handlerMappings = new ArrayList<>();
-
-    void initialize() {
-        final String basePackage = Application.class.getPackage().getName();
-        final AnnotationHandlerMapping annotationHandlerMapping = new AnnotationHandlerMapping(basePackage);
-        annotationHandlerMapping.initialize();
-        addHandlerMapping(annotationHandlerMapping);
-    }
 
     public void addHandlerMapping(final HandlerMapping handlerMapping) {
         this.handlerMappings.add(handlerMapping);
