+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/exception/InternalServerError.java
@@ -1,4 +1,4 @@
-package com.techcourse;
+package com.interface21.webmvc.servlet.mvc.exception;
 
 public class InternalServerError extends RuntimeException {
